fix(backend): align integer types and unify JSON handling for cross-backend compatibility
- Update token count fields to `BigInt`/`i64` across schema and models.
- Update cost fields to `Double`/`f64` in SQLite models.
- Wrap `serde_json::Value` with `crate::db::Json` for Diesel compatibility.
- Implement conditional compilation for connection pool interaction in `recall.rs` and `chronicle_service.rs`.
- Resolve type mismatches and trait bound errors for both PostgreSQL and SQLite backends.

+++ backend/src/models/chats.rs
@@ -1902,8 +1902,8 @@ pub struct ChatForClient {
     pub active_impersonated_character_id: Option<crate::db::DbId>,
     pub chat_mode: ChatMode,
     pub chronicle_id: Option<crate::db::DbId>, // Chronicle association (maps to player_chronicle_id in database)
-    pub total_prompt_tokens: i32,
-    pub total_completion_tokens: i32,
+    pub total_prompt_tokens: i64,
+    pub total_completion_tokens: i64,
     pub total_credits_used: crate::db::DbDecimal,
     pub total_actual_cost: crate::db::DbDecimal, // Raw API cost in dollars
     pub game_state: Option<crate::DbJson>,
